remove base bar char component
