+++ lib/components/Charts/VerticalBarChart/index.tsx
@@ -1,48 +1,96 @@
 import { ChartContainer } from "@/ui/chart"
+import { cloneDeep } from "lodash"
 import { ForwardedRef } from "react"
 import {
   Bar,
   BarChart as BarChartPrimitive,
-  CartesianGrid,
   LabelList,
   XAxis,
+  XAxisProps,
   YAxis,
+  YAxisProps,
 } from "recharts"
-
-import { xAxisProps, yAxisProps } from "../utils/elements"
+import { prepareData } from "../utils/bar"
+import {
+  xAxisProps as xAxisConfigureProps,
+  yAxisProps as yAxisConfigureProps,
+} from "../utils/elements"
 import { fixedForwardRef } from "../utils/forwardRef"
-import { prepareData } from "../utils/muncher"
 import { ChartConfig, ChartPropsBase, InferChartKeys } from "../utils/types"
 
-export type BarChartProps<
+const getMaxValueByKey = (
+  data: ({ x: unknown } & Record<string, number>)[]
+): string => {
+  const clonedData = cloneDeep(data)
+
+  let label: string = ""
+  let max: number = 0
+
+  clonedData.forEach((datapoint) => {
+    delete datapoint.x
+
+    Object.entries(datapoint as Record<string, number>).forEach(
+      ([newLabel, value]) => {
+        if (max < value) {
+          max = value
+          label = newLabel
+        }
+      }
+    )
+  })
+
+  return label
+}
+
+export type VerticalBarChartProps<
   DataConfig extends ChartConfig = ChartConfig,
   Keys extends string = InferChartKeys<DataConfig>,
 > = ChartPropsBase<DataConfig, Keys> & { label: boolean }
 
-const _BarChart = <
+const _VBarChart = <
   DataConfig extends ChartConfig,
   Keys extends string = string,
 >(
-  { dataConfig, data, xAxis, yAxis, label }: BarChartProps<DataConfig, Keys>,
+  {
+    dataConfig,
+    data,
+    xAxis,
+    yAxis,
+    label,
+  }: VerticalBarChartProps<DataConfig, Keys>,
   ref: ForwardedRef<HTMLDivElement>
 ) => {
   const bars = Object.keys(dataConfig) as Array<keyof typeof dataConfig>
+  const munchedData = prepareData(data)
+
+  const xAxisProps: XAxisProps = {
+    ...xAxisConfigureProps(xAxis),
+    type: "number",
+    dataKey: getMaxValueByKey(munchedData),
+  }
+
+  const yAxisProps: YAxisProps = {
+    ...yAxisConfigureProps(yAxis),
+    type: "category",
+    dataKey: "x",
+  }
 
   return (
     <ChartContainer config={dataConfig} ref={ref}>
       <BarChartPrimitive
+        layout="vertical"
         accessibilityLayer
         data={prepareData(data)}
         margin={{ left: 12, right: 12 }}
       >
-        <CartesianGrid vertical={false} />
-        <YAxis {...yAxisProps(yAxis)} hide={yAxis?.hide} />
-        <XAxis {...xAxisProps(yAxis)} hide={xAxis?.hide} />
+        <XAxis {...xAxisProps} hide={xAxis?.hide} />
+        <YAxis {...yAxisProps} hide={yAxis?.hide} />
 
         {bars.map((key) => {
           return (
             <>
               <Bar
+                layout="vertical"
                 key={`bar-${key}`}
                 dataKey={key}
                 fill={dataConfig[key].color}
@@ -51,7 +99,7 @@ const _BarChart = <
                 {label && (
                   <LabelList
                     key={`label-{${key}}`}
-                    position="top"
+                    position="right"
                     offset={10}
                     className="fill-foreground"
                     fontSize={12}
@@ -66,4 +114,4 @@ const _BarChart = <
   )
 }
 
-export const BarChart = fixedForwardRef(_BarChart)
+export const VerticalBarChart = fixedForwardRef(_VBarChart)
